refactor(features): merge booking into bookings (4.4)
